Add Ananke theme submodule and configuration files
This commit adds the Ananke theme submodule and the necessary configuration files to the repository. The Ananke theme is added as a submodule with the URL https://github.com/theNewDynamic/gohugo-theme-ananke.git. The following files are added or modified:
- themes/ananke (submodule)
- hugo.toml
- .gitmodules
- archetypes/default.md
- content/posts/my-first-post.md

These changes are necessary to incorporate the Ananke theme into the Hugo site and configure the basic settings.